feat(app): app folder domain design
