--- src/checkout.ts
@@ -1,13 +1,0 @@
-import { Product } from './product';
-import { Receipt } from './receipt';
-
-export class Checkout {
-
-    public scanItem(product: Product): void {
-        console.log('todo')
-    }
-
-    public generateReceipt(): Receipt {
-        return new Receipt();
-    }
-}
