--- src/product.ts
@@ -1,10 +1,0 @@
-export class Product {
-
-    public readonly name: string;
-    public readonly price: number;
-
-    constructor(name: string, price: number) {
-
-    }
-
-}
