+++ src/product/product.ts
@@ -1,0 +1,6 @@
+export class Product {
+  public readonly name: string;
+  public readonly price: number;
+
+  constructor(name: string, price: number) {}
+}
