+++ src/receipt/receipt.ts
@@ -1,0 +1,11 @@
+import { Product } from "../product/product";
+
+export interface ReceiptItem {
+  product: Product;
+  quantity: number;
+}
+
+export class Receipt {
+  public readonly items: Array<ReceiptItem> = [];
+  public readonly totalPrice: number = 0;
+}
